Add files via upload
Im working on a medicine tracker. Basically you add the medicine you take daily and add a time you take it. The app will notify you when to take the medication. So far i'm able to add the medicine, the time and date as well as a description. Id love to add repeating notifications (i.e if you need to take a medication 3 times a day so the app holds that data) and i've also had problems with user defaults, so the app wont hold the information when closed